fix tailing lines

+++ straxen/plugins/events/s2_recon_pos_diff.py
@@ -1,6 +1,7 @@
 import numpy as np
 import straxen
 import strax
+
 export, __all__ = strax.exporter()
 
 
@@ -18,40 +19,40 @@ class S2ReconPosDiff(strax.Plugin):
     save_when = strax.SaveWhen.EXPLICIT
 
     recon_alg_included = straxen.URLConfig(
-                 help='The list of all reconstruction algorithm considered.',
-                 default=('_mlp', '_gcn', '_cnn'), infer_type=False,
-                 )
+        help='The list of all reconstruction algorithm considered.',
+        default=('_mlp', '_gcn', '_cnn'), infer_type=False,
+    )
 
     def infer_dtype(self):
         dtype = [
-        ('s2_recon_avg_x', np.float32,
-         'Mean value of x for main S2'),
-        ('alt_s2_recon_avg_x', np.float32,
-         'Mean value of x for alternatice S2'),
-        ('s2_recon_avg_y', np.float32,
-         'Mean value of y for main S2'),
-        ('alt_s2_recon_avg_y', np.float32,
-         'Mean value of y for alternatice S2'),
-        ('s2_recon_pos_diff', np.float32,
-         'Reconstructed position difference for main S2'),
-        ('alt_s2_recon_pos_diff', np.float32,
-         'Reconstructed position difference for alternative S2'),
-    ]
+            ('s2_recon_avg_x', np.float32,
+             'Mean value of x for main S2'),
+            ('alt_s2_recon_avg_x', np.float32,
+             'Mean value of x for alternatice S2'),
+            ('s2_recon_avg_y', np.float32,
+             'Mean value of y for main S2'),
+            ('alt_s2_recon_avg_y', np.float32,
+             'Mean value of y for alternatice S2'),
+            ('s2_recon_pos_diff', np.float32,
+             'Reconstructed position difference for main S2'),
+            ('alt_s2_recon_pos_diff', np.float32,
+             'Reconstructed position difference for alternative S2'),
+        ]
         dtype += strax.time_fields
         return dtype
 
     def compute(self, events):
 
-        result = np.zeros(len(events), dtype = self.dtype)
+        result = np.zeros(len(events), dtype=self.dtype)
         result['time'] = events['time']
         result['endtime'] = strax.endtime(events)
         # Computing position difference
         self.compute_pos_diff(events, result)
         return result
 
-    def cal_avg_and_std(self, values, axis = 1):
-        average = np.mean(values, axis = axis)
-        std = np.std(values, axis = axis)
+    def cal_avg_and_std(self, values, axis=1):
+        average = np.mean(values, axis=axis)
+        std = np.std(values, axis=axis)
         return average, std
 
     def eval_recon(self, data, name_x_list, name_y_list):
@@ -62,7 +63,7 @@ class S2ReconPosDiff(strax.Plugin):
         # lazy fix to delete field name in array, otherwise np.mean will complain
         x_avg, x_std = self.cal_avg_and_std(np.array(data[name_x_list].tolist()))
         y_avg, y_std = self.cal_avg_and_std(np.array(data[name_y_list].tolist()))
-        r_std = np.sqrt(x_std**2 + y_std**2)
+        r_std = np.sqrt(x_std ** 2 + y_std ** 2)
         res = x_avg, y_avg, r_std
         return res
 
@@ -74,11 +75,11 @@ class S2ReconPosDiff(strax.Plugin):
             # - must exist (index != -1)
             # - must have positive AFT
             # - must contain all alg info
-            cur_s2_bool = (events[peak_type + '_index'] !=- 1)
+            cur_s2_bool = (events[peak_type + '_index'] != - 1)
             cur_s2_bool &= (events[peak_type + '_area_fraction_top'] > 0)
             for name in self.recon_alg_included:
-                cur_s2_bool &= ~np.isnan(events[peak_type+'_x'+name])
-                cur_s2_bool &= ~np.isnan(events[peak_type+'_y'+name])
+                cur_s2_bool &= ~np.isnan(events[peak_type + '_x' + name])
+                cur_s2_bool &= ~np.isnan(events[peak_type + '_y' + name])
 
             # default value is nan, it will be overwrite if the event satisfy the requirements
             result[peak_type + '_recon_pos_diff'][:] = np.nan
